+++ mithril-aggregator/src/services/signature_consumer/interface.rs
@@ -1,0 +1,12 @@
+use mithril_common::{
+    entities::{SignedEntityType, SingleSignature},
+    StdResult,
+};
+
+/// A signature consumer which blocks until messages are available.
+#[cfg_attr(test, mockall::automock)]
+#[async_trait::async_trait]
+pub trait SignatureConsumer: Sync + Send {
+    /// Returns signatures when available
+    async fn get_signatures(&self) -> StdResult<Vec<(SingleSignature, SignedEntityType)>>;
+}